MINI/requirements.txt
streamlit
opencv-python
pytesseract
pillow
pandas
google-generativeai
python-dotenv
